Refactor: Update imports and remove unused AccessibilityEnhancer
