+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import { HelmetProvider } from "react-helmet-async";
-import App from "./AppOptimized";
+import App from "./App";
 import "./index.css";
 
 // Fallback: simple passthrough provider while ThemeContext is absent
